(Implementation)Organizing_Containers_of_Balls(Algorithm)
logic:
If 'x' container stores 'a' type of balls, than 'x' must trade all the 'a' type of balls from other container at the cost of all non 'a' type balls it has. Hence sum of non 'a' balls in 'x' == sum of all 'a' balls in other container. Adding number of 'a' type in 'x' to both side. This becomes sum of xth row = ath coloumn so all we have to do is to compare if the arrays containing all horizontal sums and vertical sums are same or not.